refactor: reorganize page structure and extract reusable components
