--- src/pages/Login.tsx
@@ -1,9 +1,0 @@
-
-
-const Login = () => {
-  return (
-    <div>Login</div>
-  )
-}
-
-export default Login
