+++ src/pages/Partners/Partners.tsx
@@ -1,0 +1,26 @@
+import PartnerMarquee from '@/components/UI/PartnersMarquee';
+import HeroSection from "./components/HeroSection";
+import Banner from "./components/Banner";
+import ProgrammeOverview from "./components/ProgrammeOverview";
+import Benefit from "./components/Benefit";
+import Process from "./components/Process";
+import BecomeAPartner from "./components/BecomeAPartner";
+import BottomCTA from './components/BottomCTA';
+
+
+// ─── Page 
+
+const Partners = () => (
+  <div className="bg-white">
+    <HeroSection/>
+    <PartnerMarquee showSecondary={false} />
+    <ProgrammeOverview/>
+    <Banner />
+    <Benefit/>
+    <Process/>
+    <BecomeAPartner/>
+    <BottomCTA/>
+  </div>
+);
+
+export default Partners;
