+++ src/pages/community/Community.tsx
@@ -1,0 +1,17 @@
+import HeroSection from "./components/HeroSection";
+import Channels from "./components/Channels";
+import CommunityGuidelines from "./components/CommunityGuidelines";
+import BottomCTA from "./components/BottomCTA";
+
+const Community = () => {
+  return (
+  <>
+    <HeroSection />
+    <Channels/>
+    <CommunityGuidelines/>
+    <BottomCTA/>
+  </>
+  );
+};
+
+export default Community;
